Fixes necessary to compile on raspberry pi 4 with clang (#1055)
* Make compile on raspberry pi 4 with clang

* asm -> __asm__

* Prettyprint

+++ src/common/sha3/xkcp_sha3x4.c
@@ -70,7 +70,7 @@ static void keccak_x4_inc_reset(uint64_t *s) {
 
 static void keccak_x4_inc_absorb(uint64_t *s, uint32_t r,
                                  const uint8_t *in0, const uint8_t *in1, const uint8_t *in2, const uint8_t *in3, size_t inlen) {
-	size_t c;
+	uint64_t c;
 
 	if (s[100] && inlen + s[100] >= r) {
 		c = r - s[100];
